requirements.txt
# Design2Code Experiments - Requirements
pyyaml>=6.0
tqdm>=4.65.0
tabulate>=0.9.0
